refactor(viewer): name the expansion seam for every presentation channel, not just isolation (#3338)
`resolveIsolationIds` holds the one policy that turns raw, user-picked ids
into ids the renderer can match against a mesh. Isolation is not the only
channel with that need: hide, show and colour all match ids against mesh
ids too, and all three break on a geometry-less `IfcElementAssembly` for
the same reason isolation did.

The isolation-only name is part of the problem #3338 describes. A `hide`
handler author reading `lib/isolation/resolveIsolationIds` concludes the
module is not theirs and hand-rolls the id list, which is precisely the
"one call site every channel must remember" failure.

Pure rename plus a doc rewrite; no behaviour change. The gate's
ROUTING_MARKERS follows the new token so the eight routed channels keep
passing.
